requirements.txt
Babel==0.9.6
Flask==0.8
Flask-Babel==0.8
Flask-Login==0.1
Flask-MongoAlchemy==0.5.3
Flask-WTF==0.6
Jinja2==2.6
MongoAlchemy==0.11
WTForms==1.0.1
Werkzeug==0.8.3
amqplib==1.0.2
anyjson==0.3.1
apache-libcloud==0.9.1
celery==2.5.3
kombu==2.1.5
pymongo==2.1.1
python-dateutil==1.5
pytz==2012c
raven==1.6.0
redis==2.4.11
simplejson==2.5.0
speaklater==1.2
wsgiref==0.1.2
